+++ crates/repositories/src/profiles.rs
@@ -203,7 +203,7 @@ impl ProfilesRepository for ProfilesSqlRepository {
             })
     }
 
-    async fn stream(&self, feed_id: i64) -> Result<BoxStream<Result<StreamProfile, Error>>, Error> {
+    async fn stream(&self, feed_id: i32) -> Result<BoxStream<Result<StreamProfile, Error>>, Error> {
         profile::Entity::find()
             .select_only()
             .column(profile::Column::Id)
